requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-Discord==0.1.69
Flask-Session==0.5.0
Flask-Caching==2.1.0
python-dotenv==1.0.0
Pillow==10.2.0
gunicorn==21.2.0
eventlet==0.35.2
SQLAlchemy==2.0.25
Werkzeug==3.0.1
requests==2.31.0
python-socketio==5.11.0
python-engineio==4.8.2
discord.py==2.3.2
aiohttp==3.9.3
cachetools==5.3.2
python-slugify==8.0.4
Markdown==3.5.2
bleach==6.1.0 